Virtual Hardware - Operational PS Update
IMPORTANT: A virtual DAQ must be created using NI MAX and it must be
named 'Virtual'. If not, the DAQ will not work. It is a good idea to
match the model of the DAQ you want to emulate (PS system uses NI
PXIe-6361).

MAKE SURE YOU HAVE THE RIGHT .curve IN THE .ini FOR THE OPAS!!!! A
.curve check has not yet been implemented.

All virtual hardware is working with one or two minor bugs I've found so
far.
1) The virtual monochrometer isn't updating the position properly so the
current value won't change.
2) For some reason, the turret called is always None. I've put a bandaid
fix in v_MicroHR.py line 98, but it ignores the bigger problem.

I think the two bugs are related, but I'm not sure how.