Provide terminal feedback when a sensor is busy
Particularly because a sensor that is stuck busy at startup prevents the app from actually starting. It is nontrivial to properly fix it, but there was zero feedback to alert users as to how to fix the issue. This is a bandaid at best, but better than zero feedback, I guess.